share collecting navigation results, make more of NavigationResultRequest.Key internal
