+++ navigator/runtime-compose/api/navigator-runtime-compose.api
@@ -10,14 +10,6 @@ public final class com/freeletics/mad/navigator/compose/DialogDestination : com/
 	public synthetic fun <init> (Lkotlin/reflect/KClass;Lkotlin/jvm/functions/Function3;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
 }
 
-public final class com/freeletics/mad/navigator/compose/InitialValue$Creator : android/os/Parcelable$Creator {
-	public fun <init> ()V
-	public final fun createFromParcel (Landroid/os/Parcel;)Lcom/freeletics/mad/navigator/compose/InitialValue;
-	public synthetic fun createFromParcel (Landroid/os/Parcel;)Ljava/lang/Object;
-	public final fun newArray (I)[Lcom/freeletics/mad/navigator/compose/InitialValue;
-	public synthetic fun newArray (I)[Ljava/lang/Object;
-}
-
 public abstract interface class com/freeletics/mad/navigator/compose/NavDestination {
 }
