+++ navigator/runtime-fragment/api/navigator-runtime-fragment.api
@@ -14,14 +14,6 @@ public final class com/freeletics/mad/navigator/fragment/HandleNavigationKt {
 	public static final fun handleNavigation (Landroidx/fragment/app/Fragment;Lcom/freeletics/mad/navigator/NavEventNavigator;)V
 }
 
-public final class com/freeletics/mad/navigator/fragment/InitialValue$Creator : android/os/Parcelable$Creator {
-	public fun <init> ()V
-	public final fun createFromParcel (Landroid/os/Parcel;)Lcom/freeletics/mad/navigator/fragment/InitialValue;
-	public synthetic fun createFromParcel (Landroid/os/Parcel;)Ljava/lang/Object;
-	public final fun newArray (I)[Lcom/freeletics/mad/navigator/fragment/InitialValue;
-	public synthetic fun newArray (I)[Ljava/lang/Object;
-}
-
 public abstract interface class com/freeletics/mad/navigator/fragment/NavDestination {
 }
