+++ navigator/runtime-fragment/build.gradle
@@ -1,7 +1,6 @@
 plugins {
     alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
-    alias(libs.plugins.kotlin.parcelize)
     alias(libs.plugins.dokka)
     alias(libs.plugins.publish)
 }
@@ -59,8 +58,6 @@ dependencies {
     implementation libs.androidx.activity
     implementation libs.androidx.lifecycle.common
     implementation libs.androidx.lifecycle.runtime
-    implementation libs.androidx.viewmodel.savedstate
     implementation libs.androidx.navigation.common
     implementation libs.androidx.navigation.runtime
-    implementation libs.kotlin.parcelize
 }
