+++ navigator/runtime-fragment/src/main/java/com/freeletics/mad/navigator/fragment/HandleNavigation.kt
@@ -1,24 +1,19 @@
 package com.freeletics.mad.navigator.fragment
 
 import android.content.Context
-import android.os.Parcelable
 import androidx.activity.result.ActivityResultCaller
 import androidx.activity.result.ActivityResultLauncher
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.coroutineScope
 import androidx.navigation.fragment.findNavController
 import com.freeletics.mad.navigator.ActivityResultRequest
 import com.freeletics.mad.navigator.NavEventNavigator
-import com.freeletics.mad.navigator.NavigationResultRequest
 import com.freeletics.mad.navigator.PermissionsResultRequest
 import com.freeletics.mad.navigator.internal.AndroidXNavigationExecutor
-import com.freeletics.mad.navigator.internal.DestinationId
-import com.freeletics.mad.navigator.internal.NavigationExecutor
 import com.freeletics.mad.navigator.internal.RequestPermissionsContract
 import com.freeletics.mad.navigator.internal.collectAndHandleNavEvents
+import com.freeletics.mad.navigator.internal.collectAndHandleNavigationResults
 import kotlinx.coroutines.launch
-import kotlinx.parcelize.Parcelize
 
 /**
  * Handles the [NavEventNavigator] events while the Fragment's lifecycle is at least
@@ -36,7 +31,9 @@ public fun handleNavigation(fragment: Fragment, navigator: NavEventNavigator) {
 
     val executor = AndroidXNavigationExecutor(fragment.findNavController())
     navigator.navigationResultRequests.forEach {
-        it.registerIn(executor, lifecycle)
+        lifecycle.coroutineScope.launch {
+            executor.collectAndHandleNavigationResults(it)
+        }
     }
 
     val dispatcher = fragment.requireActivity().onBackPressedDispatcher
@@ -62,23 +59,3 @@ private fun PermissionsResultRequest.registerIn(
         handleResult(resultMap, context)
     }
 }
-
-private fun <O : Parcelable> NavigationResultRequest<O>.registerIn(
-    executor: NavigationExecutor,
-    lifecycle: Lifecycle,
-) {
-    lifecycle.coroutineScope.launch {
-        val savedStateHandle = executor.savedStateHandleFor(key.destinationId)
-        savedStateHandle.getStateFlow<Parcelable>(key.requestKey, InitialValue)
-            .collect { result ->
-                if (result != InitialValue) {
-                    @Suppress("UNCHECKED_CAST")
-                    handleResult(result as O)
-                    savedStateHandle[key.requestKey] = InitialValue
-                }
-            }
-    }
-}
-
-@Parcelize
-private object InitialValue : Parcelable
